+++ src/main/java/com/yuguanzhang/lumi/user/dto/profile/ProfileResoponseDto.java
@@ -1,4 +1,4 @@
-package com.yuguanzhang.lumi.user.dto.search;
+package com.yuguanzhang.lumi.user.dto.profile;
 
 import com.yuguanzhang.lumi.user.entity.User;
 import lombok.AllArgsConstructor;
@@ -8,12 +8,12 @@ import java.util.UUID;
 
 @Getter
 @AllArgsConstructor
-public class SearchResoponseDto {
+public class ProfileResoponseDto {
     private final UUID userId;
     private final String email;
     private final String name;
 
-    public static SearchResoponseDto searchResoponseDto(User user) {
-        return new SearchResoponseDto(user.getUserId(), user.getEmail(), user.getName());
+    public static ProfileResoponseDto searchResoponseDto(User user) {
+        return new ProfileResoponseDto(user.getUserId(), user.getEmail(), user.getName());
     }
 }
