+++ src/main/java/com/yuguanzhang/lumi/user/service/profile/ProfileService.java
@@ -1,0 +1,9 @@
+package com.yuguanzhang.lumi.user.service.profile;
+
+import com.yuguanzhang.lumi.user.dto.profile.ProfileResoponseDto;
+
+import java.util.UUID;
+
+public interface ProfileService {
+    ProfileResoponseDto getProfile(UUID userId);
+}
