+++ src/main/java/com/yuguanzhang/lumi/user/service/profile/ProfileServiceImpl.java
@@ -1,6 +1,6 @@
-package com.yuguanzhang.lumi.user.service.serach;
+package com.yuguanzhang.lumi.user.service.profile;
 
-import com.yuguanzhang.lumi.user.dto.search.SearchResoponseDto;
+import com.yuguanzhang.lumi.user.dto.profile.ProfileResoponseDto;
 import com.yuguanzhang.lumi.user.entity.User;
 import com.yuguanzhang.lumi.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -10,14 +10,14 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
-public class SearchServiceImpl implements SearchService {
+public class ProfileServiceImpl implements ProfileService {
 
     private final UserRepository userRepository;
 
     @Override
-    public SearchResoponseDto getProfile(final UUID userId) {
+    public ProfileResoponseDto getProfile(final UUID userId) {
         User user = userRepository.findById(userId)
                                   .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다."));
-        return SearchResoponseDto.searchResoponseDto(user);
+        return ProfileResoponseDto.searchResoponseDto(user);
     }
 }
