--- src/main/java/com/yuguanzhang/lumi/user/service/serach/SearchService.java
@@ -1,9 +1,0 @@
-package com.yuguanzhang.lumi.user.service.serach;
-
-import com.yuguanzhang.lumi.user.dto.search.SearchResoponseDto;
-
-import java.util.UUID;
-
-public interface SearchService {
-    SearchResoponseDto getProfile(UUID userId);
-}
